Fixed product saving bugs
Changed the moment an product gets added to the storage to match the wizard.
Added a temporaray inspect function to the storageHelper. This way you dont have to request the storage manualy.
Rewrote the product saving. Now the product gets added as JSON.
Fixed a bug where the product dont get loaded into the array.
Added an id to the products